+++ src/ClientApp.tsx
@@ -2,7 +2,7 @@ import { hydrateRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import App from './App';
 import { Provider } from 'react-redux';
-import { store } from 'state/store/Store';
+import { store } from 'state/store/store';
 
 hydrateRoot(
   document.getElementById('root') as HTMLElement,
