+++ src/components/cards/CardsList.test.tsx
@@ -1,9 +1,9 @@
 import { describe, it, vi, Mock } from 'vitest';
 import { cleanup, render, screen } from '@testing-library/react';
 import { Provider } from 'react-redux';
-import { store } from 'state/store/Store';
+import { store } from 'state/store/store';
 import { CardsList } from 'components/cards/CardsList';
-import { fetchCharacters } from 'state/reducers/CharactersReducer';
+import { fetchCharacters } from 'state/reducers/charactersReducer';
 
 describe('Cards list', () => {
   afterEach(() => {
