+++ src/components/event-card/EventCards.test.tsx
@@ -1,9 +1,9 @@
 import { describe, it } from 'vitest';
 import { cleanup, render, screen } from '@testing-library/react';
 import { Provider } from 'react-redux';
-import { store } from 'state/store/Store';
+import { store } from 'state/store/store';
 import { EventCards } from 'components/event-card/EventCards';
-import { createEvent } from 'state/reducers/EventsReducer';
+import { createEvent } from 'state/reducers/eventsReducer';
 
 describe('Event list', () => {
   afterEach(() => {
