+++ src/state/reducers/characterReducer.ts
@@ -1,7 +1,7 @@
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
 import { CardProps } from 'components/cards/Card';
 import { getCharacter } from 'services/ApiService';
-import { RootState } from 'state/store/Store';
+import { RootState } from 'state/store/store';
 
 export type CharacterGetRequestProps = {
   isLoading: boolean;
